requirements.txt
# ML & Data Science
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
imbalanced-learn==0.11.0
category-encoders==2.6.1

# XAI Libraries
shap==0.42.1
lime==0.2.0.1

# LLM Integration
openai==1.12.0
python-dotenv==1.0.0

# Web Framework
fastapi==0.103.1
uvicorn[standard]==0.23.2
jinja2==3.1.2

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Utilities
pydantic==2.3.0
loguru==0.7.0
pyyaml==6.0.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0
httpx==0.24.1
